refactor: moved project up one directory level
I moved everything in the startpage directory up one level and removed that folder. This was done so
that commitizen could work properly at all levels of the project along with yarn build/test/run
commands for local development. The alterative would be to manage two different versions of
package.json which isn't ideal.